[refactor] - Updated the package structure, added javadoc to annotations
