--- src/main/java/com/github/millij/eom/spi/IExcelBean.java
@@ -1,7 +1,0 @@
-package com.github.millij.eom.spi;
-
-/**
- * A marker interface for excel sheet mapped beans.
- */
-public interface IExcelBean {
-}
